Renamed python service wrapper to kubos_service. Added setup.py so it can be installed like any other package.
