+++ libs/kubos_service/setup.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+
+setup(name='kubos_service',
+      version='1.0',
+      packages=['kubos_service'],
+      license='MIT',
+      description='Python interface for creating kubos services',
+      )
